refactor(persistence): split E2ECheckpoint model from checkpoint.py
Move the 458-line E2ECheckpoint Pydantic model and exception classes
to persistence/_checkpoint_model.py, bringing checkpoint.py from 1010
lines down to 551 (under the 600-line ceiling).

checkpoint.py re-exports CheckpointError, ConfigMismatchError, and
E2ECheckpoint for full backward compatibility — no callers need updating.
